% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{List of Google Trends categories}
\format{
A data frame with 1426 rows and 2 variables
}
\usage{
data(categories)
}
\description{
A data frame containing a list of the names and id values of all valid Google Trends categories.
The variables are as follows:
}
\details{
\itemize{
\item name. Names of valid Google Trends categories
\item id. Ids of valif Google Trends categories
}
}
\keyword{datasets}
