% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphAverages.R
\name{getGraphAverages}
\alias{getGraphAverages}
\title{Extracts Google Trends average search interest data associated with multiple topics}
\usage{
getGraphAverages(
  terms,
  geo = NULL,
  property = NULL,
  category = NULL,
  startDate = NULL,
  endDate = NULL,
  api.key
)
}
\arguments{
\item{terms}{The search terms of interest. A minimum of two and Up to five search terms can be queried simultaneously.}

\item{geo}{Applies a geographical restriction to data query. Parameter should be a string, please refer to ISO-3166-2 for the supported values.}

\item{property}{Defines the Google property to query data from. Options are web/images/news/froogle(Shopping)/youtube. Parameter should be a string, defaults to 'web' when unspecified.}

\item{category}{Defines a category filter for data extraction. Please consult the Trends Explore page for possible legal input here. Parameter should be a string, defaults to 'All categories' when unspecified.}

\item{startDate}{Defines the starting date for data extraction. Start date should be a month and a year in the format YYYY-MM e.g. 2010-01. Parameter should be a string, defaults to '2004-01' when unspecified.}

\item{endDate}{Defines the end date for data extraction. End date should be a month and a year in the format YYYY-MM e.g. 2010-01. Parameter should be a string, defaults to the current month and year when unspecified.}

\item{api.key}{Google API access key.}
}
\value{
Returns the average search volume of a Google Trends time series.
}
\description{
Extracts Google Trends average search interest data associated with multiple topics
}
\examples{
\dontrun{
getGraphAverages(c("apple","orange"), api.key = "YOUR_API_KEY")

getGraphAverages(c("Toyota","Fiat","Chevrolet"), geo = "US", property = "news",
category = "Car Video", startDate = "2014-01", endDate = "2014-12", api.key = "YOUR_API_KEY")
}
}
