% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegions.R
\name{getRegions}
\alias{getRegions}
\title{Extracts Google Trends search interest data per country}
\usage{
getRegions(
  terms,
  geo = NULL,
  property = NULL,
  category = NULL,
  startDate = NULL,
  endDate = NULL,
  api.key
)
}
\arguments{
\item{terms}{The search term of interest. Only a single term can be queried at a time.}

\item{geo}{Applies a geographical restriction to data query. Parameter should be a string, please refer to ISO-3166-2 for the supported values.}

\item{property}{Defines the Google property to query data from. Options are images/news/froogle(Shopping)/youtube. Parameter should be a string, defaults to 'web' when unspecified.}

\item{category}{Defines a category filter for data extraction. Please consult with the Trends Explore page for possible legal input here. Parameter should be a string, defaults to 'All Categories' when unspecified.}

\item{startDate}{Defines the starting date for data extraction. Start date should be a month and a year in the format YYYY-MM e.g. 2010-01. Parameter should be a string, defaults to '2004-01' when unspecified.}

\item{endDate}{Defines the end date for data extraction. End date should be a month and a year in the format YYYY-MM e.g. 2010-01. Parameter should be a string, defaults to the current month and year when unspecified.}

\item{api.key}{Google API access key.}
}
\value{
Extracts country-level search interest data, scaled in relation to the country with the higher proportion of searches for the term or topics of interest.
}
\description{
Extracts Google Trends search interest data per country
}
\examples{
\dontrun{
getRegions("apple", api.key = "YOUR_API_KEY")

getRegions(c("apple","pear","orange"), geo = "US", property = "news",
category = "Food & Drink", startDate = "2014-01", endDate = "2014-12",
api.key = "YOUR_API_KEY")
}
}
