% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopQueries.R
\name{getTopQueries}
\alias{getTopQueries}
\title{Extracts information about the top queries associated with the requested search term}
\usage{
getTopQueries(
  terms,
  geo = NULL,
  property = NULL,
  category = NULL,
  startDate = NULL,
  endDate = NULL,
  api.key
)
}
\arguments{
\item{terms}{The search term of interest. Only a single term can be queried at a time.}

\item{geo}{Applies a geographical restriction to data query. Parameter should be a string, please refer to ISO-3166-2 for the supported values.}

\item{property}{Defines the Google property to query data from. Options are images/news/froogle(Shopping)/youtube. Parameter should be a string, defaults to 'web' when empty.}

\item{category}{Defines a category filter for data extraction. Please consult with the Trends Explore page for possible legal input here. Parameter should be a string, defaults to 'All Categories' when empty.}

\item{startDate}{Defines the starting date for data extraction. Start date should be a month and a year in the format YYYY-MM e.g. 2010-01. Parameter should be a string, defaults to '2004-01'.}

\item{endDate}{Defines the end date for data extraction. End date should be a month and a year in the format YYYY-MM e.g. 2010-01. Parameter should be a string, defaults to the current month and year.}

\item{api.key}{Google API access key.}
}
\value{
Returns the top 25 queries associated with the provided search term or topic.
}
\description{
Extracts information about the top queries associated with the requested search term
}
\examples{
\dontrun{
getTopQueries("apple", api.key = "YOUR_API_KEY")

getTopQueries(terms="Amazon", geo = "US", property = "news",
startDate = "2014-01", endDate = "2014-12", api.key = "YOUR_API_KEY")
}
}
