% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrendsAPI-package.R
\docType{package}
\name{gtrendsAPI-package}
\alias{gtrendsAPI}
\alias{gtrendsAPI-package}
\title{gtrendsAPI: Accessing the official Google Trends API}
\description{
gTrendsAPI provides a platform to access data from the official Google Trends API through R.
}
\author{
\strong{Maintainer}: Ricardo Correia \email{rahc85@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
