% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\docType{data}
\name{properties}
\alias{properties}
\title{List of Google Trends properties}
\format{
A list with 4 valid properties
}
\usage{
data(properties)
}
\description{
A list containing the names of valid Google Trends properties.
}
\keyword{datasets}
