% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{List of Google Trends regions}
\format{
A data frame with 3380 rows and 2 variables
}
\usage{
data(regions)
}
\description{
A data frame containing a list of valid Google Trends regions.
The variables are as follows:
}
\details{
\itemize{
\item regionCode. Region ISO code
\item regionName. Name of the region
}
}
\keyword{datasets}
